Day 10 Part 2: Use an upper bound on the solution to limit nodes
I tried this earlier and for some reason it works now.
I think I was missing the check for whether the subproblem was feasible
(the second if).
I stole it from the child generation of `Node`.

Using the upper bound seems to help a lot but there are still cases in
the real input that we run out of memory on.