Day 12: Use vector of Bool instead of Int
To represent that a `tile` x,y was part of a present at a given
coordinate, we previously used that (linearised) coordinate in the
assertion.
This had the potential downside that `Int`s are much less restrained
than `Bool`s and some of the logic was a little weird.
A tile is now a vector of 9 bools, where at most can be true. These
bools represent which of the 9 squares of a present this tile is.
This hopefully makes the search space smaller / more well-defined?

We also now print the solution found, which is pretty fun.